requirements.txt
openai
dotenv
jupyter
ipykernel
azure-search-documents
azure-identity
pypdf
langchain
pandas
